+++ code/qcommon/huffman.c
@@ -328,14 +328,14 @@ static void Huff_transmit (huff_t *huff, qint ch, byte *fout) {
 
 void Huff_Decompress(msg_t *mbuf, qint offset) {
 	qint			ch, cch, i, j, size;
-	byte		seq[65536] = {0};
+	byte		seq[MAX_INFO_STRING*2] = {0};
 	byte*		buffer;
 	huff_t		huff;
 
 	size = mbuf->cursize - offset;
 	buffer = mbuf->data + offset;
 
-	if ( size <= 0 ) {
+	if ( size < 2 ) {
 		return;
 	}
 
@@ -351,6 +351,9 @@ void Huff_Decompress(msg_t *mbuf, qint offset) {
 	// don't overflow with bad messages
 	if ( cch > mbuf->maxsize - offset ) {
 		cch = mbuf->maxsize - offset;
+	}
+	if ( cch > sizeof( seq ) ) {
+		cch = sizeof( seq );
 	}
 	bloc = 16;
 
@@ -369,7 +372,7 @@ void Huff_Decompress(msg_t *mbuf, qint offset) {
 				ch = (ch<<1) + get_bit(buffer);
 			}
 		}
-    
+
 		seq[j] = ch;									/* Write symbol */
 
 		Huff_addRef(&huff, (byte)ch);								/* Increment node */
@@ -380,15 +383,19 @@ void Huff_Decompress(msg_t *mbuf, qint offset) {
 
 void Huff_Compress(msg_t *mbuf, qint offset) {
 	qint			i, ch, size;
-	byte		seq[65536];
+	// worst compression ratio is ~1.2x (rounded up to 2x) plus 2 plus 256 of possible NYT's
+	byte		seq[MAX_INFO_STRING*4 + 2 + 256];
 	byte*		buffer;
 	huff_t		huff;
 
 	size = mbuf->cursize - offset;
 	buffer = mbuf->data + offset;
 
-	if (size<=0) {
+	if ( size <= 0 ) {
 		return;
+	}
+	if ( size > MAX_INFO_STRING*2 ) {
+		size = MAX_INFO_STRING*2;
 	}
 
 	Com_Memset(&huff, 0, sizeof(huff_t));